Make all os-collect-config polling distro agnostic
Ubuntu does not support journalctl (which we currently use
in toci_test.sh to poll for os-collect-config to finish).

This patch wraps our os-collect-config polling commands so that distro
specific commands can be used to poll for the events.

NOTE: eventually it would be cool to be able to poll for specific
heat events, etc.

Change-Id: I15db5ab57f32c5f365a702a4a6d20e2debc427bd