Updates to work with latest setup-baremetal cmds.
TOCI updates to work with the latest triple-incubator
scripts.

Recently we changed setup-baremetal and register-nodes in
the incubator so that the way it handles MACS works differently.
Specifically we now only pass the MAC addresses we want
register-nodes to actually use (previously that script contained
some rather janky logic to decide which MACs to use on its own).
This makes good sense for the devtest story and makes testing
w/ VMs more sensible.

When using TOCI w/ real bare metal (using a predined set of MAC
addresses and power management commands) we make the following
assumptions:

 1) when setting up the seed VM we assign the first MAC in the list

 2) when setting up the undercloud we skip the first MAC and assign
   all the others.

** This should work well and allow us to bring up a single undercloud
node and possibly a multi-node overcloud as well.

This patch also includes updates to pass in parameters to
setup-baremetal instead of relying on environment variables for
everything.

Lastly, this change also gets rid of some TOCI logic to handle
using real bare metal nodes (using power management commands etc).
Now that we have support in incubator/register-nodes for power
management we can just use that instead.

Change-Id: I112c4a18c8e9d6e8ad5f152305ff018e79e97ecc