Add element to install a testenv worker
This element adds a new script testenv-worker, who purpose is to register
itself with a gearman broker. When called it holds the test environment
lock while it calls back to the client using a gearman callback job. The
callback job is passed data describing the test environment. Once finished
the original job is completed, freeing up the environment.

This is intended to be used by tripleo-ci to discover and lock test
environments.

Depends on I170ee12955263065a4adff4221e5bc60015cae70

Change-Id: I79323cb78edfcea7a7d56d9cd507354ea420ee65