Add script to install a testenv client
This commit adds a new script testenv-client, who purpose is to call
a gearman function "lockenv" passing it the name of a unique callback
gearman job, the callback when called can be used to run a command
while the test environment remains in a locked state.

This is intended to be used by tripleo-ci to lock test environments while
being used.

Change-Id: I5d552a1a8571f6ce0514c535f064e238d37dbc1f