Add "ci_commands" ssh command filter
ci_commands is added to roots authorized_keys as
command="/usr/local/bin/ci_commands on the test hosts. And the
private key is being passed to the ci slaves. The slaves can now
use this key to ssh to the test environment and run a subset of
commands(filtered by ci_commands).

Change-Id: If08b81b8ff5ba6ce526406962e2c76097d141cf5
Co-Authored-by: Chris Jones <cmsj@tenshu.net>
Co-Authored-by: Robert Collins <rbtcollins@hp.com>