Run overcloud deploy tests against the seed.
Currently we deploy the overcloud on the undercloud, but there is no
conceptual benefit to this - the undercloud is meant to be tested
entirely via the undercloud test; the seed is the same code as the
undercloud, differing only in the use of static configuration rather
than heat delivered configuration. Avoiding the undercloud
build-and-deploy makes the overcloud test substantially faster.

Change-Id: Id5895be09ad6728f9d83d260371d02777ee217e8