Reduce size of log tarball's
The binary systemd logs files don't zip well, compressing with xz at a
compression level of 3 matches the speed of compressing with gzip but
halfs the space used. Also filtering out /etc/services and /etc/pki as
these tend to take up a lot of space. This reduces the Fedora tarballs
from 2MB to about 700KB.

Change-Id: Id0d5d3b9587d2739bf9197f8b3a13549df00ae6f