testenv-worker: use signals to enable/disable jobs
Updates the testenv-worker script to handle signals so that
we can stop/start the processing of new jobs without
killing the existing testenv-worker process. This is
important in that it will give us better control of
testenv-worker node upgrades without disrupting the
job queue.

The following signal handling is now supported:

 USR1: Enable job processing.
 USR2: Disable job processing. The current job will finish processing
       and no new jobs will be accepted. The testenv-worker process
       will however stay running.
 INT:  Exit after any currently running job finishes. The testenv-worker
       process must be restarted after this signal is received in
       order to process new jobs.

Change-Id: I465f34393627f6b6fd6584824531d01287b523d7