Use /opt/git instead of /opt/stack/new
The generalized repo caching mechanism in nodepool puts copies of
every repo in gerrit into /opt/git, which is an improvement over
the selective caching of some repos into ~jenkins/workspace-cache.
To support a cleaner move to a diskimage-builder version of
nodepool, migrate to consuming repos from /opt/git instead of
from /opt/stack/new.

/opt/git on nodepool images is owned by root.

This patch needs a change to the tripleO jobs to land (or a
similar patch to allow the jenkins user access to /opt/git)
in order to work.

Depends-On: Iebe3b5f194682e759266553b0f1e861f2958b048
Change-Id: Id35c2b99a889f2af2baa10351be21e136e2c5cd0