Use devtest_testenv default when building CI TE
Resources allocated to our CI testenvs should match the current
default that developers are using. So use the same defaults when
building a testenv. This will result in the testenv defining nodes
with 2G of memory instead of 4G.

Also export NODE_ARCH, without doing this it would be ignored.

Change-Id: I1ee2a9ce661822cfca24c407607fa82f479efc16