Add revert hook point
This can be used to temporarily revert commits in other repositories to
allow CI to be unblocked. A bug number should be included, this is then
used to decide if the revert is to be skipped (if the bug number passed
into temprevert is mentioned in the commit that triggers CI the revert
will be skipped).

Change-Id: Iee47b4ecdecd6435dea3e2ebb2981cef1ecac881