Add a load test script - bighammer.sh
This script can be used to give some basic metrics about the
reliability of a cloud allowing us to make an informed decision about
its suitability for CI. As is parts of it could be improved but this can
be an exercise when burning in the next cloud.

It has been used in order to assess hp1 suitability to being added back
actively running CI jobs.

Change-Id: I239bfdfcabe0e3fe531c2751e9d1583b2bb9ee4d