Add /mnt/state/var/log contents to indexed test logs.
This patch expands the files we extract from our instance tarballs.
This is to get these logs indexed by logstash, which expands troubleshooting
and analysis capabilities.

Extracting contents of /mnt/state/var/log to $instance/mnt, if that
directory exists in our tarball.  No mnt dir created if not.

Change-Id: I71a0080d68aeb25d8254783d20e28dcce6f95482
Partial-Bug: #1328645