Explicitly set DIB_REPOREF_*
Explicitly tell source-repositories which revision of a project to pull
in, source-repositories by default pulls in "master", which is ok for
all projects except the one we actually want to test (devstack-gate
leaves the commit we want checked out in a detached head).

Upto now we haven't been testing the correct version of most of the
commits in the experimental queue.

Change-Id: Ibb4be09b22179d62c9d6f0afb2525c521bb3f36d
Closes-Bug: #1364345