Cherry pick Nova fix for extra_specs bug
Cherry pick a Nova fix that resolves a scheduling error due to
the fact that extra_specs data was being canonicalized in the
Ironic driver but not in the relevant Nova scheduler filters.

Also, removes the previous Ironic temprevert that is no
longer needed (the Ironic drivers now subclass the Nova ones).

Change-Id: I45f3e0401709435ccf96d69e1703a880d40b8660