Set tripleo ci control variables based on job name
Currently if we change any of these variables we wont know if we did the
right thing until is merged into openstack-infra/config. Moving these
into tripleo-ci will give us the ability to make changes and have CI
validate we did the correct thing. It will also ensure all jobs run with
the same defaults (e.g. currently not all the jobs run with
USE_CIRROS=1 which I don't think was ever intended).

Finally we will soon be turning back on hp1 which has a different test
network IP range, so we will need the ability to set the IPs of the
proxy and geard server based on rack, which can be done here not in
openstack-infra/config.

Change-Id: I565bb10442ce4eb73fc91397d22dd1fce25cd35b