Update virsh ssh commands to support setting the locale
The patch https://review.openstack.org/#/c/124061 in Ironic is changing
the SSH commands to force the commands to use their default language
(whatever language the strings in the application was written in, for
virsh that's english) for output by passing LC_ALL=C as part of the
command line. This is needed because as we do some pattern matchings
with the output of some commands and if the system is configured with
a non-english locale setting the matching is going to fail.

This patch makes the locale variable optional when checking the virsh
commands so it would work with the old and new commands.

Change-Id: I7b6bf0c93c88888231d260036c765228e57b848f