Force install of mariadb.i686 1:5.5.39-1.fc20
A Fedora update to this package is causing pip to fail to build
the mysql extension.

Until we get the problem fixed pull in a patch that forces
dib to install the old version. Doing this in tripleo-image-elements
would also work but needs to be done in multiple places since this is
for ci only doing it here should be fine.

Related-Bug: #1401957
Change-Id: Iba2e52d2ce0e5c25c1ad7945bd4f8f1622432be6