Debugging aids for failed tar unpacking
We have had quite a few CI failures where we can't diagnose the cause of
the failure because g_s_f_h fails to unpack the tar file containing the
logs.

The intent of this change is to gather a little more information that
might help us figure out why we weren't able to unpack the tar.

Partial-bug: 1414098

Change-Id: Idb9115885295572d5afb825b814959b45ff429df