Disentangle tarring in logfiles and increase logging
We're having problems with some part of creating this tarfile; but
trying to create the tarfile and move it across the network all in one
step is making it difficult to figure out what's going wrong.

This change is intended to be slower and more methodical. Instead of
doing this all at once, we use several SSH commands to create a tempdir;
then to create a tarfile inside it; then to scp that file onto our
working machine; then to unpack it.

This will probably be slower, but should give us more information about
how each step is proceeding - which should in turn enable us to figure
out what's breaking and what we can do about it.

Also increasing our ssh log level to Verbose so we see relevant error
messages when our ssh/scp commands fail.

Change-Id: I867869bb516ea70312eaf5f00ec071ad652c44e8
Partial-bug: 1414098