Add support for Fedora 21 image downloads
The Fedora 21 image doesn't have any updates so our existing
TOCI download script doesn't work there yet. This patch
updates the Fedora download logic such that if an updated
image exists it is used. Otherwise we download the Fedora
release image from cloud.fedoraproject.org.

DIB_RELEASE is now required to be explicitly set when using the
new function so that was defaulted to Fedora 20 (for now).

Change-Id: Ib3ea92ae17884251d86888e5f035f67025a155fa