Set root disk == node disk for puppet jobs
When using Puppet we don't want an ephemeral partition.
This patch configures the Puppet CI job such that the
ROOT_DISK is equal to the default NODE_DISK size
which means Ironic won't waste space on the ephemeral
partition.

This change will be important (required even) as we
expand the CI make use of things like Ceph which uses
/srv/data for its data storage by default... it would
be very easy to exceed the space on the root partition.

By simply making full use of the root disk we allow
project defaults to work seamlessly with our TripleO CI.

Change-Id: Icd85ed156fe167573cad976772c32bca1c84a04a