Switch overcloud-f21-nonha to instack
Move overcloud-f21-nonha from testing the bash elements and instead make use of
instack. This currently involves a number of projects that have to get pulled
into the upstream openstack infrastructure. There are also some workarounds in
here that we will work on removing while various dependecies move around.

This commit breaks tripleo-ci into 2 distinct code paths once all of the jobs
are ported to instack we should be able drop the legacy code path(along with
support for tripleo-incubator) and return to a single code path.

This ci builds centos7 images eventually the instack based CI to support both
CentOS and Fedora.

Co-Authored-By: Dan Prince <dprince@redhat.com>

Depends-On: I1acd19da8567ab93b5003caf67673cc70efea5fa
Change-Id: I09207c1f8db4e9c0fde92a254bc03fb4e6276f11