Delorean go faster stripes
Ideally we would like to build packages inline so that we elliminate
the overhead setting up build images and user a new container for
each package build. To get to this point we need to be running jobs
on a distro that matches the images we are testing.

Until we get to this point there are a number of thing we can do, primarily
o Downlod a prebuild delorean/centos image
o Start a single building container and just reuse it for each package
o Pin the dolerean build container to a spcific cenots mirror

Change-Id: I94984d069884a254a037eae3cd2d309909213721