Pull tripleo related packages from delorean/current
Pulling from current means we will no longer need to build all
of the packages, instead we would only need to build whatever ZUUL
tells us to build. Because we are no longer building  several
packages in the same ci run we can get off our TripleO fork of
delorean as the speed difference will be a lot less.

This commit also includes changes to handle the rename of
ironic-discoverd to ironic-inspector as both sets of changes
are now needed to unblock CI.

Co-Authored-By: John Trowbridge <trown@redhat.com>

Change-Id: Ia879942910c367dc47942f2e0c63351abaf16f63
Depends-on: Id4855d9f4275912440f197f900a59fe5e3bea693
Depends-on: I319f818deeeafc36968b73b8d249c5bbcaa0d2ea