Set puppet DIB_INSTALLTYPE in tripleo.sh not toci_instack.sh
Currently we override the puppet modules to build from source for CI
but that leads to tripleo.sh being broken for developers wanting to
set up a local environment matching what's in CI.

So, instead rely on tripleo.sh to setup the right version based on
the branch (e.g master or stable/liberty based on STABLE_RELEASE)

We still set up puppet.env so the DIB_REPOREF data is setup correctly
when we depend on an under-review puppet change.

This depends on the https://review.openstack.org/#/c/246447/ change to
tripleo.sh (merged for master tripleo-common), not using
Depends-On or this change will be blocked on the (not merged) liberty
patch.

Change-Id: Ia7615d7f6293e4a220e0121dbb5ea095b3bb5275