Set overcloud deploy timeout
This timeout defaults to 4 hours, which is longer than the gate will
allow our jobs to run, which means that if something in the deploy
hangs we get killed by the gate before we have a chance to collect
logs.  This makes debugging such problems next to impossible, so
let's set a sane timeout that is shorter than the gate timeout to
give us a chance to collect logs before we get killed.

Change-Id: I17ab0414ba80b0ced7233b9bce7a4191bb682b34