Wait for pacemaker cluster to settle
When using pacemaker in the HA job, there is no guarantee that all the
resources are started and running and the cluster is settled before we
run the overcloud ping test. This can lead to various errors, a common
one being 503 Service Unavailable when we try to create the tenant-stack
in Heat during the pingtest, since often Heat is one of the last
resources to start in the cluster.

Change-Id: I794fad76744d2a9a63122a3391c33830220b8ce6
Partial-Bug: #1545318