Roll up static Heat envs into CI directory
As we've started testing more features we have grown some static
Heat environments that are specific to CI.  Instead of generating
these inline in the CI scripts, let's just check them into the
repo as standalone files and copy them to the undercloud for its
use.

Note that although there are only two such files right now, I will
almost certainly need more for testing overcloud SSL, and it seems
likely that as TripleO adds features we will have a need to
add more test environment files to CI.

I've also reordered this commit before the network isolation one
that will use the net-iso.yaml file because I think we need to do
some additional configuration for that and it makes more sense to
just do it in the standalone files right away.

Change-Id: Ia73db99c24637b8c9ca02848141597e8ecfa3881