Test overcloud SSL
For the moment, I am enabling this on the nonha job even though the
ha job is where network isolation is enabled.  This is for two
reasons: 1) Network isolation in CI is not working yet so it won't
be possible to get a useful run out of it, and 2) We seem to assign
the VIPs consistently enough in CI that I think we can get away
with not having network isolation for SSL testing.

Change-Id: I4943c6c74e0be96c1d7e190908b9262df05d059a