IPv4 network isolation testing for HA and Ceph
This patch adds network isolation on the ceph and HA CI
jobs using the t-h-t multiple-nics templates. We recently
rebuilt our test environments so that each VM has 8 NICs
(only 6 are required for this, 2 extra). The multiple-nics
templates will use one nic per isolated network.

This patch wires in custom code so that the undercloud (seed vm)
acts as the gateway for the external network. No custom routing
should be required. Additionally, we supply custom parameters
to tripleo.sh when using NETISO_V4 such that custom 10. ranges
are used for the private and floating networks.

Co-Authored-By: Giulio Fidente <gfidente@redhat.com>
Co-Authored-By: Derek Higgins <derekh@redhat.com>

Change-Id: I830755d7d882d155813c105dff76900e6675cca2
Depends-On: I22c4d37ce4f489105a89d3039a0689ad1b4b8c9f