Add common bash functions to help track metrics.
Add a bash library that has several functions that can
be used within TOCI to track simple metrics (and eventually)
load them into a graphite server as time series data.

The record_metric function can be used to store any metric
in the configured data file.

The start_metric/stop_metric functions are used to store and
calculate the total time a task tasks. These functions can be
called in a nested fashion such that timestamps within timestamps
can be tracked and recorded.

The metrics_to_graphite function is meant to be called after all metrics
have been recorded. It will also reset the data so it isn't posted
to graphite twice (should you choose to batch this).

Change-Id: I35228ec61b4c6c468b277ff48bf57872a5b24362