Deploy ceph in the nonha job
The ceph job has been renamed to "upgrades", to continue testing
ceph support we need to add it to the nonha job. There is another
patch Ic0196596df6a11b02333d06f63b2e13c137c0bcc that is going to
be adding upgrades support to toci.

As is, this patch essentially will leave the upgrades job to a nonha
job without ceph.

Depends-On: I6110b71e484af749838f91dc5c6c4982b0c83074
Change-Id: I508610432efcbcf88f065a01f20b805adf7ce912