Hide the postci logs
If ci fails early in the run and as a results some resources
are not ready for postci to access them then error messages are
generated that have nothing to do with the root cause of the
problem. Pipe postci error messages to a file so that they don't
get confused with the actual error.

Change-Id: I9ca394d17d76f701d7a4fe0f19f61fd0747cf46d