Test overcloud deletion in periodic job
Recently I've seen a number of breakages where Heat can't delete a
stack that was deployed with network isolation enabled.  To avoid
promoting a repo with a broken Heat, we should test it in the
periodic job.

I'm not adding it to any of the regular jobs because the stack
deletion takes a significant amount of time (3.5 minutes in one
job I checked) and the main thing I'm concerned with is not
promoting a broken Heat.

Change-Id: I23ff8e9b2cc3676957a13d7efc42831693a97afe