Use split cirros image for ping test
Since cirros is so much lighter weight in a nested virt environment,
this should help with the load on our CI systems.

Note that we switched away from Cirros originally due to the
no_timer_check bug that is present in it, but when the split image
with a separate kernel and initrd is used Nova has some magic that
adds the necessary kernel param automatically[1].  That should make
this safe to use again.

It also allows us to move back to the m1.tiny flavor and use a much
smaller volume, which should make the ping test even faster.
Deleting a 10 GB volume in particular is painful because Cinder
zeros out the storage space before it reports the deletion as
complete.

Change-Id: I5081c00a1cc9a191172e6c32b68c66c110d18c15
1: https://github.com/openstack/nova/blob/ce5f9c609f7938c5f80839dd86dfcb33ad796b1c/nova/virt/libvirt/driver.py#L4090