Dump end of logs to console.log on failure
The image build and undercloud install output are both sent to log
files to keep the size of console.log manageable.  However, this
also means we can't write logstash queries against failures in
those steps because logstash can only see console.log.  This change
adds a tail of the log file when the logged command fails so we at
least have a chance of capturing the error in console.log.

Also cleans up the log location messages.  This file...this file
where both refer to different files is just confusing.

Change-Id: I8d5cf3a19b44a2f53934f3c22c01d82406252fd0