Actually test stack-update to new tht
Modifies the upgrades CI job to actually test stack-update'ing to the
tripleo-heat-template patch being tested by CI.

On the initial overcloud deploy, openstack-tripleo-heat-templates from
delorean-current or delorean (for stable branches) will be used. On the
stack-update, the version from the delorean-ci repo will be used instead
(if one was built, otherwise it will just install from delorean-current
or delorean again).

Change-Id: Ifb1e62cc6ca824261fba7b2e190c627d8b537450