Add final success/failure message to ping test
The actual result of the ping test is about halfway through the
output of the test, and the DONE message at the end is not all
that meaningful since the test should never not complete.  This
change replaces that message with a more specific one based on the
exitval variable.

Change-Id: Ib6deeb87b975c87c3269958ad88c0d0991b42828