Switch to a local dns server
We're seeing lots of failures with DNS requests in CI jobs,
it looks like the problem is happening outside of our overcloud
deployment. This is particularly bad for DNS requests presumably
because its UDP. So switch our primary DNS server to point at named
on the bastion server, while we try and find the root cause.

Change-Id: Ic59e338e4d3abee15c6d6614dd377ca81479850b