Re-write tripleo-jobs to use gerrit
Per openstack-infra:

http://lists.openstack.org/pipermail/openstack-dev/2016-April/092482.html

We no longer have direct access to the Jenkins logs. This patch
re-writes the CI script to use Gerrit instead of Jenkins for the
CI results. We will maintain the gerrit version of this script
alongside of the old one for now to see if Jenkin's access opens
back up in the future.

Co-Authored-By: Carlos Camacho <ccamacho@redhat.com>

Change-Id: I9852838a92a857bb576ed7337aacf5cf617da66a