Don't use cached undercloud image for element repos either
These repos install packages on the undercloud, and if a package
install is mistakenly removed/broken it won't be caught if we use
the image that already has the packages installed.  This actually
happened in Ie204c5d0aa0f48d3d93d834b51a644d8562cd2fd which broke
all undercloud installs not using the cached image.

Change-Id: Ife3392f9c66616fb91c208fc75c38ee99246072b