Don't remove packages from the delorean-deps repository
To allow the toci scripts to run twice on the same machine we were
removing all packages installed from a delorean repository. This
no longer works as delorean-deps now contains python-six and removing
it fails because yum depends on it. Instead only remove from the
delorean repositories that contain openstack packages.

Change-Id: I4c80fa7a91768747dcb5d1d8f7020f01ec78f626