Fail pingtest more quickly
This patch introduces two changes with the intent of failing faster on
the pingtest. The first is to not enter the wait_for if the stack-create
operation fails immediately. This can happen if the pingtest template
contains invalid data.  The second is to add a fail short-circuit
argument to wait_for for CREATE_FAILED. Previously, we wait around for
CREATE_SUCCESS even if the stack creation has already failed.

Change-Id: I49d81572d2eb1a53ee66dc545ae261de3d0795c9