Make stack deletion work on all branches
Piping yes to heat isn't working properly - we get 141 return
values from the calls, which indicates a pipe failure.  Upon
further review, we can do openstack stack delete --yes on every
branch except liberty, so let's just do that.  Then we don't need
to pipe anything because liberty doesn't have the delete
confirmation functionality.

The wait_for also doesn't work on liberty because the openstack
stack command doesn't exist yet there.  The openstack function
doesn't help because the command is embedded in wait_for.  Instead,
we just need to pass the correct command to wait_for.

Since there's only a couple of instances of this and the split
between heatclient and osc is temporary I'm just copying the logic
where it needs to go.  Once we're all osc it can be removed along
with the openstack function itself.

Change-Id: I31f880c7fccd55fcd233256f3fc1f2fa63024be3