Sync up EndpointMap in enable-tls with default enable-tls from THT
We need to change the public_url endpoint to use IP_ADDRESS instead
of CLOUDNAME in CI so we use a custom enable-tls file. This needs
to be updated and kept in sync with the default TLS map we ship
in tripleo-heat-templates though as CI won't be testing the version
from THT.

A sample case is change I0cab3cdb2189dab3844f2eda52b8697d05ad3447
which fixed the port where the Sahara public_url is announced
but we didn't fix this in the tripleo-ci version of enable-tls.yaml
causing failures in CI only.

Closes-Bug: 1583151

Change-Id: I48a09de05c4d07853faef1959d67e8f109b50b49