Add mulitnode CI job support to tripleo-ci
This would allow using multinode jobs spawned by nodepool, but that is
not the only use case. Any cloud provider could be used by this setup,
including a 3rd party CI setup.

The scripts here are modified to support installing the undercloud
directly on the node running the toci_gate_test.sh script, which is
often the jenkins slave. When deploying an all in one Overcloud this
results in needing 2 nodes total to do the basic TripleO deployment.

There is a new script named toci_instack_multinode.sh to run the
multinode setup.

Change-Id: Icc67a289df207348536c0a2419b3fc81f03fe758