Configure a different port for httpd ipxe
The default port for httpd ipxe in CI is 8088, which conflicts with
the port used for the telnet server that streams the job console log.
Configure the undercloud to use a different port (3816) instead to
avoid the conflict. When a different port is configured, we no longer
have to kill the telnet server.

Change-Id: Iaa67444ff4d65f628a5ab1b384d2f246669ceaa5