Skip pingtest cleanup
We're spending around 2 minutes per run cleaning up the pingtest
resources in an environment that will be deleted immediately
afterwards.  While there's some small chance of this finding a bug
in the overcloud services, we're deploying a very simple Heat stack
so IMHO the slim chances of this finding a bug are not worth the
time it's taking.

Change-Id: I9b39bb6070baa956db283f5a7691385173290e6d