Delete ping test environment in periodic jobs
I9b39bb6070baa956db283f5a7691385173290e6d removed the deletion of
the ping test stack and resources to save time on CI jobs.  However,
because we only have Heat convergence enabled in the overcloud that
means we may miss convergence delete bugs.  To avoid promoting a
repo with such a bug, let's run the ping test cleanup in the
periodic jobs where runtime is less of a concern.

Change-Id: I55672297b63952d011675071b24aa9894316d1c1