Add --yes flags to heat stack-delete for liberty
My tests are failing because they are waiting for a manual
confirmation on heat stack-delete:

 heat stack-delete tenant-stack
Are you sure you want to delete this stack(s) [y/N]?

heat-manage shows we are on liberty: heat-manage --version
5.0.1

So first check if the --yes flag is available under that
liberty deployment. If it is, add the --yes flag, if not
present continue without confirmation.

Change-Id: I2b5b69715879a195f34785d995f1a4aed0c49ab3