Temporary fix for nodejs/libuv conflict
Temporarily remove epel-release, and don't enable epel via tripleo.sh.
Also remove nodejs and nodejs-devel. This is a quick fix for:
https://bugs.launchpad.net/tripleo/+bug/1606685

We have a CI dependency on EPEL too with moreutils, so that needs
to be installed before we remove epel.  Since everything is running
on the Jenkins slave that should be okay for now.

The proper fix will be to remove epel more completely going forward:
https://review.openstack.org/347499
https://review.openstack.org/347513

Co-Authored-By: Ben Nemec <bnemec@redhat.com>
Change-Id: I91eff35586547dd278e8a32668f06efb840fc8c6
Partial-Bug: #1606685