Remove Ceph from the ha job
Now that we've re-enabled the nonha job we're testing ceph there,
so we can remove it from the ha job and reduce the number of nodes
needed by one.  This should also speed up the ha job.  Also, at the
moment we aren't testing the non-ceph path at all in our check jobs
which is not good either.

Also switches the nonha job to use storage-environment.yaml so we
keep coverage of that file.  It turns out there's no practical
difference between that and puppet-ceph-devel now because the
composable work removed the deployment of osds on controllers.

Change-Id: I67a4a6f9348e181204d5ee5afe377b8e65066dc1