scenario002: deploy Swift
Add Swift to scenario002 and switch Glance backend to use Swift storage.

Note: scenario002 now looks very similar to regular multinode job, this
is going to change as we'll add more services in scenario002. Now we
keep the non-ha multinode job to be run all the time, since scenarios
are only run at demand when Zuul thinks we need it.

Change-Id: I29ed91bef3ee822453b064ee1fd28e4549a4e130