Use current-tripleo for undercloud upgrades
When testing the undercloud major upgrade, we should be upgrading to
current-tripleo (rdo-trunk-master-tripleo) insetad of
rdo-trunk-master-tested.

$STABLE_RELEASE is unset in deploy.sh before running the upgrade. The
undercloud_upgrade function in tripleo.sh can then make use of
repo_setup to enable the new repositories instead of having to redo
those steps manually.

The python-cachetools workaround is removed since that fix has now
merged and the TripleO pin has moved to include the fix as well.

Change-Id: Ie79ba6d6f3fa235eb76f6fc1957b7b2caffc812c