Rate limit testenv creation
Adds a named semaphore to the testenv-worker code that will
prevent more than 5 testenvs from creating at once.  This requires
the use of the posix_ipc Python module, which is added to the
te-broker puppet manifest.

Right now if someone pushes a patch series of tripleo-ci changes,
which each start 8 jobs at the moment, we can end up with 20 or 30
testenvs creating concurrently.  Not surprisingly, this does not
work well and most of those testenvs will fail or timeout.  Using
an approximate testenv creation time of 3 minutes (which is based
solely on my observation of how long the Heat stacks generally take),
the 5 concurrent creations allowed should process up to 30 testenvs
in the 20 minute timeout of the te-broker.  Even if we do exceed
that number, the testenvs that _do_ get processed are far more
likely to succeed in a less loaded environment so our overall pass
rate will be higher than all 30+ at once.  And of course we can
increase the concurrency if we find that 5 at once can be handled
easily, which would only increase the potential throughput.

The one major drawback of this approach is that Linux IPC is fairly
terrible at error-handling, so if a testenv-worker process holding
a semaphore dies in some ungraceful fashion that doesn't allow the
Python code to release the semaphore, we may have to reboot the
te-broker to clean up.  I don't anticipate such a situation happening
often in the simple te-broker environment, but it's something to be
aware of.

Change-Id: Id80105a1578aa2120d2508018e53846affe254a0