Add configurable mtu to --mulitnode-setup
Testing multinode environments on different clouds has shown that mtu on
the configured openvswitch interfaces may need to be lower than the
hardcoded value used in ovs_vxlan_bridge function from devstack-gate,
which is 1450.

This patch adds a $MTU variable which can be set to reconfigure the mtu
on the created interfaces after ovs_vxlan_bridge has been run.

This patch also makes a small change to make $BASE overrideable (which
is a variable used by devstack-gate).

Change-Id: I89147bb732664ba143446baa1c69232370db98bb