Source deploy.env in tripleo.sh
When running multinode jobs, we need to make sure deploy.env is sourced
everytime tripleo.sh is run. The easiest way to do this is to source it
at the beginning of tripleo.sh if the file exists in the expected
location where it would be created by tripleo-ci.

We then need to pass the value over ssh when we ssh to subnodes to run
commands.

Change-Id: I01bf15b3c689b4fd0cdd848cf264dd9794e1d828
Closes-Bug: #1634319