Update ovb baremetal flavor to match infra flavor
We've increased the baremetal flavor to have 8 GB of RAM and 4
vcpus in order to help with resource starvation issues in CI
jobs.  This just updates the prepare script to match.

This also requires increasing the nodepool quota.  Some quick math
suggests this would now max out about 2.7 TB, so I'm setting the
quota to about 3 TB so there's a little head room for things like
orphaned stacks and infra vms.

Change-Id: I3d354c2e5f76e220d07895bc2bc262eabe5ebb88