Add worker config envs in toci_gate_test
Because tripleo.sh now sources deploy.env on every run, we can't
make changes to deploy args in deploy.sh.  Instead we can add the
appropriate envs to the deploy args in toci_gate_test.  This also
means we need to regenerate the update args with the new deploy
args.

Change-Id: Ibc0e577b138c712d3fd4bbc5186b73cc33d244bd
Partial-Bug: 1642429