Enable consuming packages for a feature branch
An use case for TripleO CI can be to test changes on an
specific feature branch. In that case, the packages need to be
consumed for that specific branch, instead of master.
This change enables the detection of a feature branch in the
current change, and enables the related repositories (they will
default to master if the feature branch for a repo does not exist)

Also update ELEMENTS_PATH properly to consune latest
features from diskimage-builder

Change-Id: I701c01dd18023c0c05aede9d06b6b1380b6f2386
Needed-By: I0ca6a4ae3a2684d473b44e5f332ee4225ee30f8c