Properly set distro branch in DLRN when STABLE_RELEASE=newton
RDO changed the distgit branch for stable releases in newton, using
newton-rdo instead of rpm-newton. This was done to unify specs
between RDO Trunk and final versions, but Tripleo CI was not updated
to match.

Without this patch, DLRN will try to use a non-existing rpm-newton
branch for Netwon stable releases, always defaulting to rpm-master.

Change-Id: I4268dd2b4814d0d223fdb897cccfaf950567d0cd